@@ -82,8 +82,9 @@
     "disc": "^1.3.2",
     "envify": "^4.1.0",
     "es6-promise": "^4.1.1",
-    "eslint": "^4.3.0",
-    "eslint-config-sanity": "^2.1.4",
+    "eslint": "^5.9.0",
+    "eslint-config-prettier": "^3.3.0",
+    "eslint-config-sanity": "^0.140.0",
     "get-uri": "^2.0.1",
     "gzip-size": "^3.0.0",
     "karma": "^1.7.0",
@@ -97,6 +98,7 @@
     "lodash.once": "^4.1.1",
     "mocha": "^3.4.2",
     "nyc": "^11.0.3",
+    "prettier": "^1.15.2",
     "pretty-bytes": "^4.0.2",
     "rimraf": "^2.6.1",
     "uglify-js": "^3.0.26",
